Hide `useLayoutEffect` warning when using SSR
When using SSR and invoked `useLayoutEffect` inside the component, a warning is prompted.

```
Warning: useLayoutEffect does nothing on the server, because its effect cannot be encoded into the server renderer's output format. This will lead to a mismatch between the initial, non-hydrated UI and the intended UI. To avoid this, useLayoutEffect should only be used in components that render exclusively on the client. See https://reactjs.org/link/uselayouteffect-ssr for common fixes.
```

Please see the [hack](https://medium.com/@alexandereardon/uselayouteffect-and-ssr-192986cdcf7a) proposed by @alexandereardon.